fix counter tests

+++ repository/src/main/java/com/softserve/edu/device/test/data/BbiDeviceTestData.java
@@ -119,8 +119,8 @@ public class BbiDeviceTestData implements DeviceTestData {
     }
 
     @Override
-    public int getInitialCapacity() {
-        return (int)getLongValue("initialCapacity");
+    public long getInitialCapacity() {
+        return getLongValue("initialCapacity");
     }
 
     @Override
